pyproject.toml
[project]
name = "gojeera"
version = "0.1.0"
description = "A TUI for interacting with Atlassian Jira from your terminal."
readme = "README.md"
requires-python = ">=3.10,<3.15"

authors = [
    { name = "Gaston T.", email = "pythondeveloper@outlook.com" },
    { name = "Vadim Khitrin", email = "me@vkhitrin.com" },
]
maintainers = [{ name = "Vadim Khitrin", email = "me@vkhitrin.com" }]
keywords = ["tui", "jira", "cli", "terminal", "shell", "task-management"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Framework :: Textual",
    "Framework :: Pydantic :: 2",
    "Framework :: Pytest",
    "Framework :: Sphinx",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Terminals",
    "Topic :: Utilities",
]
dependencies = [
    "atlas-doc-parser",
    "click==8.3.0",
    "defusedxml>=0.7.1",
    "gitpython>=3.1.45",
    "httpx>=0.28.1",
    "mdit-py-plugins>=0.5.0",
    "pydantic-settings[yaml]>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-json-logger>=4.0.0",
    "python-magic>=0.4.27",
    "textual-autocomplete>=4.0.6",
    "textual-fspicker",
    "textual-image>=0.8.4",
    "textual-jumper",
    "textual-tags",
    "textual[syntax]>=6.4.0",
    "xdg-base-dirs>=6.0.2",
]

[project.urls]
Homepage = "https://github.com/vkhitrin/gojeera"
Documentation = "https://github.com/vkhitrin/gojeera/tree/main/docs"
Repository = "https://github.com/vkhitrin/gojeera"
Issues = "https://github.com/vkhitrin/gojeera/issues"

[dependency-groups]
dev = ["textual-dev>=1.8.0"]
lint = ["ruff>=0.14.3", "ty>=0.0.13", "vulture"]
test = [
    "pytest>8.4.1",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>6.2.1",
    "respx>=0.22.0",
]

[project.scripts]
gojeera = "gojeera.cli:gojeeraCLI"

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.mypy]
python_version = "3.14"
strict_optional = false
ignore_missing_imports = true
show_error_context = false
exclude = ["docs", "test"]

[tool.uv.sources]
atlas-doc-parser = { git = "https://github.com/MacHu-GWU/atlas_doc_parser-project" }
textual-tags = { git = "https://github.com/Zaloog/textual-tags" }
textual-jumper = { git = "https://github.com/Zaloog/textual-jumper", rev = "main" }
textual-fspicker = { git = "https://github.com/davep/textual-fspicker" }
vulture = { git = "https://github.com/jendrikseipp/vulture" }

[tool.vulture]
paths = ["src"]
ignore_names = [
    "BINDINGS",
    "DEFAULT_CSS",
    "ENVIRONMENT",
    "FIX_VERSIONS",
    "HELP",
    "SPRINT",
    "STORY_POINTS",
    "TEAM",
    "VERSIONS",
    "_align_to_target",
    "_handle_account_id_ready",
    "_on_*",
    "_selected_work_item_type_id",
    "_stored_title",
    "_token_to_content",
    "absolute_offset",
    "action_*",
    "action_focus_content",
    "action_focus_previous_or_tabs",
    "action_jump_to_first",
    "action_jump_to_last",
    "action_new_work_item_subtask",
    "action_show_overlay",
    "apply_completion",
    "as_dict",
    "as_json",
    "check_action",
    "clear_form",
    "clear_information",
    "cls",
    "compact",
    "content_scroll",
    "current_status_color",
    "date_metadata_widgets_container",
    "discover",
    "displayed_count",
    "end",
    "file_secret_settings",
    "get_candidates",
    "get_matches",
    "get_search_string",
    "handle_*",
    "has_pending_changes",
    "is_loading",
    "jump_mode",
    "markup",
    "model_config",
    "on_*",
    "open_file_picker",
    "padding",
    "pagination",
    "parameters",
    "progress",
    "raw",
    "relation_type",
    "result_subtitle",
    "selected_user",
    "settings_customise_sources",
    "should_show_dropdown",
    "start",
    "start_at",
    "static_fields_container",
    "sync_client",
    "template",
    "tooltip",
    "unified_search_button",
    "unified_search_input",
    "unified_search_mode_selector",
    "validate_*",
    "visible",
    "watch_*",
    "work_item_components_has_changed",
    "work_item_due_date_has_changed",
    "work_item_parent_has_changed",
    "work_item_search_results",
    "work_item_time_tracking",
    "work_log_items_container",
]
